whatpm/t/tree-test-1.dat

#data
<!Doctype html><body>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE XHTML><title></title><body>
#errors
not HTML5
#document
| <!DOCTYPE XHTML>
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML5><title></title><body>
#errors
not HTML5
#document
| <!DOCTYPE HTML5>
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE html4><title></title><body>
#errors
not HTML5
#document
| <!DOCTYPE html4>
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML 2.0//EN"><title></title><body>
#errors
not HTML5 (quirks doctype)
#document
| <!DOCTYPE HTML "-//IETF//DTD HTML 2.0//EN" "">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "HTML"><title></title><body>
#errors
not HTML5 (quirks doctype)
#document
| <!DOCTYPE HTML "HTML" "">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"><title></title><body>
#errors
not HTML5 (quirks doctype)
#document
| <!DOCTYPE HTML "-//W3C//DTD HTML 4.01 Transitional//EN" "">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"
    "http://www.w3.org/TR/html4/loose.dtd"><title></title><body>
#errors
not HTML5 (limited quirks doctype)
#document
| <!DOCTYPE HTML "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Frameset//EN"><title></title><body>
#errors
not HTML5 (quirks doctype)
#document
| <!DOCTYPE HTML "-//W3C//DTD HTML 4.01 Frameset//EN" "">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Frameset//EN"
    "http://www.w3.org/TR/html4/frameset.dtd"><title></title><body>
#errors
not HTML5 (limited quirks doctype)
#document
| <!DOCTYPE HTML "-//W3C//DTD HTML 4.01 Frameset//EN" "http://www.w3.org/TR/html4/frameset.dtd">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"><title></title><body>
#errors
not HTML5 (limited quirks doctype)
#document
| <!DOCTYPE HTML "-//W3C//DTD XHTML 1.0 Transitional//EN" "">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
    "http://www.w3.org/TR/html4/loose.dtd"><title></title><body>
#errors
not HTML5 (limited quirks doctype)
#document
| <!DOCTYPE HTML "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD XHTML 1.0 Frameset//EN"><title></title><body>
#errors
not HTML5 (limited quirks doctype)
#document
| <!DOCTYPE HTML "-//W3C//DTD XHTML 1.0 Frameset//EN" "">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD XHTML 1.0 Frameset//EN"
    "http://www.w3.org/TR/html4/loose.dtd"><title></title><body>
#errors
not HTML5 (limited quirks doctype)
#document
| <!DOCTYPE HTML "-//W3C//DTD XHTML 1.0 Frameset//EN" "http://www.w3.org/TR/html4/loose.dtd">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "HTML5"
    "http://www.w3.org/TR/html4/loose.dtd"><title></title><body>
#errors
not HTML5 (no quirks doctype)
#document
| <!DOCTYPE HTML "HTML5" "http://www.w3.org/TR/html4/loose.dtd">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML SYSTEM "http://www.ibm.com/data/dtd/v11/ibmxhtml1-transitional.dtd"><title></title><body>
#errors
not HTML5 (quirks doctype)
#document
| <!DOCTYPE HTML "" "http://www.ibm.com/data/dtd/v11/ibmxhtml1-transitional.dtd">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML><!DOCTYPE HTML><title></title><body>
#errors
duplicate DOCTYPE
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML><!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"><title></title><body>
#errors
duplicate DOCTYPE
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN"><!DOCTYPE HTML><title></title><body>
#errors
not HTML5
duplicate DOCTYPE
#document
| <!DOCTYPE HTML "-//W3C//DTD HTML 4.01 Transitional//EN" "">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "XSLT-compat"><title></title>
#errors
#shoulds
XSLT-compat
#document
| <!DOCTYPE HTML "XSLT-compat" "">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "xslt-compat"><title></title>
#errors
not HTML5
#shoulds
#document
| <!DOCTYPE HTML "xslt-compat" "">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC " XSLT-compat"><title></title>
#errors
not HTML5
#shoulds
#document
| <!DOCTYPE HTML " XSLT-compat" "">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "XSLT-compat  "><title></title>
#errors
not HTML5
#shoulds
#document
| <!DOCTYPE HTML "XSLT-compat  " "">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "XSLT-generated"><title></title>
#errors
not HTML5
#document
| <!DOCTYPE HTML "XSLT-generated" "">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "xslt-compat"><title></title>
#errors
not HTML5
#document
| <!DOCTYPE HTML "xslt-compat" "">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML PUBLIC "XSLT-compat" "http://example.com/"><title></title>
#errors
not HTML5
#document
| <!DOCTYPE HTML "XSLT-compat" "http://example.com/">
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE NONHTML PUBLIC "XSLT-compat"><title></title>
#errors
not HTML5
#document
| <!DOCTYPE NONHTML "XSLT-compat" "">
| <html>
|   <head>
|     <title>
|   <body>

#data
a<html>
#errors
no doctype
<html> in body
#document
| <html>
|   <head>
|   <body>
|     "a"

#data
a<html a>
#errors
no doctype
<html> in body
#document
| <html>
|   a=""
|   <head>
|   <body>
|     "a"

#data
<p>x<html a>y
#errors
no doctype
<html> in body
#document
| <html>
|   a=""
|   <head>
|   <body>
|     <p>
|       "xy"

#data
<!DOCTYPE HTML><html> <head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html> 
<head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html>   <!---->   <head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <!--  -->
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html>  <!DOCTYPE HTML>  <head><body>
#errors
<!DOCTYPE> in <html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html a>  <html b>  <head><body>
#errors
second <html>
#document
| <!DOCTYPE HTML>
| <html>
|   a=""
|   b=""
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html>   </end>  <head><body>
#errors
unmatched </end>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html title="t"><head><title></title></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   title="t"
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML><html manifest=manifest.dat><title></title><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   manifest="manifest.dat"
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML><html><head title="t"><title></title></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     title="t"
|     <title>
|   <body>

#data
<!DOCTYPE HTML><html><head><title class="t"></title></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|       class="t"
|   <body>

#data
<!DOCTYPE HTML><html><head></head><body><title class="a"></title></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <title>
|       class="a"

#data
<!DOCTYPE HTML><html><head></head><body></body><title class="a"></title></html>
#errors
data after body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <title>
|       class="a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body title="t"></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     title="t"

#data
<!DOCTYPE HTML><html><head><title></title><meta name="a"></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <meta>
|       name="a"
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title></head><body><meta name="a"></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <meta>
|       name="a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body><meta name="a"></html>
#errors
data after body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <meta>
|       name="a"

#data
<!DOCTYPE HTML><html><head><title></title><link rel="a"></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <link>
|       rel="a"
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title></head><body><link rel="a"></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <link>
|       rel="a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body><link rel="a"></html>
#errors
data after body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <link>
|       rel="a"

#data
<!DOCTYPE HTML><html><head><title></title><base href="http://"></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <base>
|       href="http://"
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title></head><body><base href="http://"></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <base>
|       href="http://"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body><base href="http://"></html>
#errors
data after body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <base>
|       href="http://"

#data
<!DOCTYPE HTML><p><b><p>X
#errors
<b> is not closed before second <p>
<b> is not closed before end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|     <p>
|       <b>
|         "X"

#data
<!DOCTYPE HTML><b><p>
#errors
<b> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <b>
|       <p>

#data
<!DOCTYPE HTML><b><p>X
#errors
<b> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <b>
|       <p>
|         "X"

#data
<!DOCTYPE HTML><ins><p>XX</ins>YYY
#errors
1,32,not closed:p
1,34,not closed:ins
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ins>
|       <p>
|         "XXYYY"

## when script is disabled
#data
<!DOCTYPE HTML><noscript>aaa<noscript>bbb</noscript>ccc</noscript>
#errors
in noscript:#character
unmacthed noscript end tag
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>
|     "aaa"
|     <noscript>
|       "bbb"
|     "ccc"

## when script is disabled
#data
<!DOCTYPE HTML><body><noscript>aaa<noscript>bbb</noscript>ccc</noscript>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noscript>
|       "aaa"
|       <noscript>
|         "bbb"
|       "ccc"

## when script is disabled
#data
<!DOCTYPE HTML><head><noscript><noscript>
#errors
<noscript> in noscript
no </noscript>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>

## when script is disabled
#data
<!DOCTYPE HTML><head><noscript><noscript> 
#errors
<noscript> in noscript
no </noscript>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|       " "
|   <body>

## when script is disabled
#data
<!DOCTYPE HTML><head><noscript> <noscript> 
#errors
<noscript> in noscript
no </noscript>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|       "  "
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript><base><title></title><body>
#errors
<base> in <noscript>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|     <base>
|     <title>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript><title></title><body>
#errors
<title> in <noscript> in <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|     <title>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head></head><noscript></noscript>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noscript>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript><script></script>
#errors
<script> in <noscript> in <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|     <script>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript><body>
#errors
<body> in <noscript> in <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript><frameset></frameset>
#errors
<frameset> in <noscript> in <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <frameset>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript><br>
#errors
<br> in <noscript> in <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>
|     <br>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript><p>
#errors
<p> in <noscript> in <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>
|     <p>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript></p>
#errors
</p> in <noscript> in <head>
unmatched </p>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript></br>
#errors
</br> in <noscript> in <head>
</br> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript></div>
#errors
</div> in <noscript> in <head>
unexpected end of file (no </noscript>)
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript></img>
#errors
</img> in <noscript> in <head>
unexpected end of file (no </noscript>)
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript></head><link><body>
#errors
</head> in <noscript> in <head>
<link> after <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|     <link>
|   <body>

#data
<!DOCTYPE HTML><head></head></head><link><body>
#errors
</head> after </head>
<link> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <link>
|   <body>

#data
<!DOCTYPE HTML></noscript><head><body>
#errors
</noscript> before <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><head></noscript><link><body>
#errors
</noscript> in <head>
<link> in <body>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <link>

#data
<!DOCTYPE HTML><head></head></noscript><link><body>
#errors
</noscript> after </head>
<link> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <link>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript></body><meta></noscript><link><body>
#errors
</body> in <noscript> in <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|       <meta>
|     <link>
|   <body>

## When scripting is disabled
#data
<!DOCTYPE HTML><head><noscript></html><meta></noscript><link><body>
#errors
</html> in <noscript> in <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|       <meta>
|     <link>
|   <body>

#data
<!DOCTYPE HTML></p>
#errors
</p> before <head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML></br>
#errors
</br> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML></head></p>
#errors
</p> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><head></body><link>x
#errors
</body> in head
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <link>
|   <body>
|     "x"

#data
<head></html><link>
#errors
no <!DOCTYPE>
</html> in <head>
#document
| <html>
|   <head>
|     <link>
|   <body>

#data
<head></p><link>
#errors
no <!DOCTYPE>
</p> in head
#document
| <html>
|   <head>
|     <link>
|   <body>

#data
<head></head></p>
#errors
no <!DOCTYPE>
</p> after </head>
#document
| <html>
|   <head>
|   <body>

#data
</p>
#errors
no <!DOCTYPE>
</p> before </head>
#document
| <html>
|   <head>
|   <body>

#data
<head></head></div>
#errors
no <!DOCTYPE>
</div> after </head>
#document
| <html>
|   <head>
|   <body>

#data
<head></head></unknown>
#errors
no <!DOCTYPE>
</unknown> after </head>
#document
| <html>
|   <head>
|   <body>

#data
<head></head></head>
#errors
no <!DOCTYPE>
</head> after </head>
#document
| <html>
|   <head>
|   <body>

#data
<head></head></html>x
#errors
no <!DOCTYPE>
</html> after </head>
#document
| <html>
|   <head>
|   <body>
|     "x"

#data
<head></head></body><body>y</body>
#errors
no <!DOCTYPE>
</body> after </head>
#document
| <html>
|   <head>
|   <body>
|     "y"

#data
<head></head></noscript>
#errors
no <!DOCTYPE>
</noscript> after </head>
#document
| <html>
|   <head>
|   <body>

#data
<head></head></style>
#errors
no <!DOCTYPE>
</style> after </head>
#document
| <html>
|   <head>
|   <body>

#data
<head></br><link>
#errors
no <!DOCTYPE>
</br> in </head>
<link> in </head>
#document
| <html>
|   <head>
|     <link>
|   <body>

#data
<!DOCTYPE HTML><head></head></img>
#errors
</img>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><head></head></div>
#errors
unmatched </div>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><head></head></body>
#errors
</body> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><head></head></html>
#errors
</html> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><head><title></title></head><base><body>
#errors
<base> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <base>
|   <body>

#data
<!DOCTYPE HTML><head><title></title></head><meta><body>
#errors
<meta> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <meta>
|   <body>

#data
<!DOCTYPE HTML><head></head><title></title><body>
#errors
<title> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML><iframe>
#errors
no </iframe>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <iframe>

#data
<!DOCTYPE HTML><iframe></iframe>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <iframe>

#data
<!DOCTYPE HTML><iframe>a</iframe>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <iframe>
|       "a"

#data
<!DOCTYPE HTML><noframes>
#errors
no </noframes>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noframes>
|   <body>

#data
<!DOCTYPE HTML><noframes></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noframes>
|   <body>

#data
<!DOCTYPE HTML><noframes>a</noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noframes>
|       "a"
|   <body>

#data
<!DOCTYPE HTML><noframes>
#errors
no </noframes>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noframes>
|   <body>

#data
<!DOCTYPE HTML><noframes></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noframes>
|   <body>

#data
<!DOCTYPE HTML><xmp>
#errors
no </xmp>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <xmp>

#data
<!DOCTYPE HTML><xmp></xmp>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <xmp>

#data
<!DOCTYPE HTML><xmp>a</xmp>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <xmp>
|       "a"

#data
<!DOCTYPE HTML><html><head><title></title><style></style></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <style>
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title><style type="text/css"></style></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <style>
|       type="text/css"
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title><style type="text/css">p {}</style></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|     <style>
|       type="text/css"
|       "p {}"
|   <body>

#data
<!DOCTYPE HTML><html><style type="text/css"></style><title></title></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <style>
|       type="text/css"
|     <title>
|   <body>

#data
<!DOCTYPE HTML><style type="text/css"></style><title></title><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <style>
|       type="text/css"
|     <title>
|   <body>

#data
<!DOCTYPE HTML><title></title><body><p>a<style type="text/css"></style>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <p>
|       "a"
|       <style>
|         type="text/css"

#data
<!DOCTYPE HTML><form><form></form>
#errors
form in form
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>

#data
<!DOCTYPE HTML><ul><li>a<li>c</ul>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         "a"
|       <li>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><ul><li>a<div>b<li>c</ul>d
#errors
div closed by li
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         "a"
|         <div>
|           "b"
|       <li>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dt>a<dt>c</dl>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         "a"
|       <dt>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dt>a<dd>c</dl>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         "a"
|       <dd>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dd>a<dt>c</dl>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "a"
|       <dt>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dd>a<dd>c</dl>d
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "a"
|       <dd>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dt>a<div>b<dt>c</dl>d
#errors
div closed by dt
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         "a"
|         <div>
|           "b"
|       <dt>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dt>a<div>b<dd>c</dl>d
#errors
div closed by dd
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         "a"
|         <div>
|           "b"
|       <dd>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dd>a<div>b<dt>c</dl>d
#errors
div closed by dt
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "a"
|         <div>
|           "b"
|       <dt>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><dl><dd>a<div>b<dd>c</dl>d
#errors
div closed by dd
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "a"
|         <div>
|           "b"
|       <dd>
|         "c"
|     "d"

#data
<!DOCTYPE HTML><body><form><div>a</form>b
#errors
no </div> before </form>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <div>
|         "a"
|     "b"

#data
<!DOCTYPE HTML><body><form><div><div><div>a</form>b
#errors
no </div> before </form>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <div>
|         <div>
|           <div>
|             "a"
|     "b"

#data
<!DOCTYPE HTML><form><p></form>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <p>

#data
<!DOCTYPE HTML><form><p></form>x
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <p>
|     "x"

#data
<!DOCTYPE HTML><form><li></form>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <li>

#data
<!DOCTYPE HTML><form><li></form>y
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <li>
|     "y"

#data
<!DOCTYPE HTML><form><dt></form>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <dt>

#data
<!DOCTYPE HTML><form><dt></form>x
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <dt>
|     "x"

#data
<!DOCTYPE HTML><form><dd></form>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <dd>

#data
<!DOCTYPE HTML><form><dd></form>x
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <dd>
|     "x"

#data
<!DOCTYPE HTML><ul></body>
#errors
ul not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>

#data
<!DOCTYPE HTML><ul><li></body>
#errors
ul not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>

#data
<!DOCTYPE HTML><dl></body>
#errors
dl not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>

#data
<!DOCTYPE HTML><dl><dt></body>
#errors
dl not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>

#data
<!DOCTYPE HTML><dl><dd></body>
#errors
dl not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>

#data
<!DOCTYPE HTML><p></body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>

#data
<!DOCTYPE HTML><table>  </table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       "  "

#data
<!DOCTYPE HTML><table>  a</table>
#errors
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <table>
|       "  "

#data
<!DOCTYPE HTML><table>  a b</table>
#errors
char in <table>
x
3
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a b"
|     <table>
|       "  "

#data
<!DOCTYPE HTML><table>a b </table>
#errors
char in <table>
x
x
4
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a b "
|     <table>

#data
<!DOCTYPE HTML><table><br> </table>
#errors
<br> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <br>
|     " "
|     <table>

#data
<!DOCTYPE HTML><table><tbody>a </table>
#errors
char in <tbody>
x2
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a "
|     <table>
|       <tbody>

#data
<!DOCTYPE HTML><table><tr>a </table>
#errors
char in <tr>
x2
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a "
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table>a<tr> </table>
#errors
char in <table>
char in <tr>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     " "
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody><br> </table>
#errors
<br> in <tbody>
char in <tbody>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <br>
|     " "
|     <table>
|       <tbody>

#data
<!DOCTYPE HTML><table><tr><br> </table>
#errors
<br> in <tr>
char in <tr>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <br>
|     " "
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table a><table b>  </table></table>
#errors
<table> in <table>
unmatched </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|     <table>
|       b=""
|       "  "

#data
<!DOCTYPE HTML><table a><table b>  </table> </table>
#errors
<table> in <table>
unmatched </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|     <table>
|       b=""
|       "  "
|     " "

#data
<!DOCTYPE HTML><table><tr><td><table>  </table></td>x</table>
#errors
char in <tr>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "x"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <table>
|               "  "

#data
<!DOCTYPE HTML><table><tr><td></td>x<table>  </table></table>
#errors
char in <tr>
no </table> before <table>
unmatched </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "x"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|     <table>
|       "  "

#data
<!DOCTYPE HTML><table><tr><td><table>x</table><table> </table></table>
#errors
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "x"
|             <table>
|             <table>
|               " "

#data
<!DOCTYPE HTML><table><title>a</title></table>
#errors
<title> in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <title>
|       "a"
|     <table>

#data
<!DOCTYPE HTML><table><li><table></table>
#errors
<li> in <table>
<table> in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <li>
|     <table>
|     <table>

#data
<!DOCTYPE HTML><table><input> </table>
#errors
<input> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <input>
|     " "
|     <table>

#data
<!DOCTYPE HTML><table><input type="text"> </table>
#errors
<input> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <input>
|       type="text"
|     " "
|     <table>

#data
<!DOCTYPE HTML><table><input type=hidden> </table>
#errors
<input type=hidden> in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <input>
|         type="hidden"
|       " "

#data
<!DOCTYPE HTML><table> <input> </table>
#errors
<input> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <input>
|     " "
|     <table>
|       " "

#data
<!DOCTYPE HTML><table>a<input> </table>
#errors
char in <table>
<input> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <input>
|     " "
|     <table>

#data
<!DOCTYPE HTML><table> <input type="text"> </table>
#errors
<input> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <input>
|       type="text"
|     " "
|     <table>
|       " "

#data
<!DOCTYPE HTML><table>a<input type=text> </table>
#errors
char in <table>
<input> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <input>
|       type="text"
|     " "
|     <table>

#data
<!DOCTYPE HTML><table> <input type=hidden> </table>
#errors
<input type=hidden> in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       " "
|       <input>
|         type="hidden"
|       " "

#data
<!DOCTYPE HTML><table>a<input type=hidden> </table>
#errors
char in <table>
<input> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <input>
|       type="hidden"
|     " "
|     <table>

#data
<!DOCTYPE HTML><table><input type="HiddeN"> </table>
#errors
<input type=hidden> in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <input>
|         type="HiddeN"
|       " "

#data
<!DOCTYPE HTML><table><thead><input type=hidden></table>
#errors
<input type=hidden> in <thead>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <input>
|           type="hidden"

#data
<!DOCTYPE HTML><table><thead>a<input type=hidden></table>
#errors
char in <thead>
<input type=hidden> in <thead>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <input>
|       type="hidden"
|     <table>
|       <thead>

#data
<!DOCTYPE HTML><table><tr><input type=hidden></table>
#errors
<input type=hidden> in <tr>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <input>
|             type="hidden"

#data
<!DOCTYPE HTML><table><tr>a<input type=hidden></table>
#errors
char in <tr>
<input type=hidden> in <tr>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <input>
|       type="hidden"
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><script></script></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <script>

#data
<!DOCTYPE HTML><table> <script></script> </table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       " "
|       <script>
|       " "

#data
<!DOCTYPE HTML><table>a<script></script> </table>
#errors
char in <table>
<script> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <script>
|     " "
|     <table>

#data
<!DOCTYPE HTML><table><tfoot><script></script> </table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|         <script>
|         " "

#data
<!DOCTYPE HTML><table><tfoot>a<script></script> </table>
#errors
char in <tfoot>
<script> in <tfoot>
char in <tfoot>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <script>
|     " "
|     <table>
|       <tfoot>

#data
<!DOCTYPE HTML><table><tr> <script></script> </table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           " "
|           <script>
|           " "

#data
<!DOCTYPE HTML><table><tr>a<script></script> </table>
#errors
char in <tr>
<script> in <tr>
char in <tr>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <script>
|     " "
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><style></style></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <style>

#data
<!DOCTYPE HTML><table> <style></style> </table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       " "
|       <style>
|       " "

#data
<!DOCTYPE HTML><table>a<style></style> </table>
#errors
char in <table>
<style> in <table>
char in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <style>
|     " "
|     <table>

#data
<!DOCTYPE HTML><table><tfoot><style></style> </table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|         <style>
|         " "

#data
<!DOCTYPE HTML><table><tfoot>a<style></style> </table>
#errors
char in <tfoot>
<style> in <tfoot>
char in <tfoot>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <style>
|     " "
|     <table>
|       <tfoot>

#data
<!DOCTYPE HTML><table><tr> <style></style> </table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           " "
|           <style>
|           " "

#data
<!DOCTYPE HTML><table><tr>a<style></style> </table>
#errors
char in <tr>
<style> in <tr>
char in <tr>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <style>
|     " "
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><b>x<i>y<div>z</b>a</i>b
#errors
1,26,in table:b
1,27,in table:#character
1,30,in table:i
1,31,in table:#character
1,36,in table:div
1,37,in table:#character
1,41,in table:/b
1,41,not closed:div
1,42,in table:#character
1,46,in table:/i
1,46,not closed:div
1,46,in table:#character
no </div></table> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <b>
|       "x"
|       <i>
|         "y"
|     <i>
|     <div>
|       <i>
|         <b>
|           "z"
|         "a"
|       "b"
|     <table>

#data
<!DOCTYPE HTML><table><tr><td></body>
#errors
unmatched </body>
table not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<!DOCTYPE HTML><table><tr><th></body>
#errors
unmatched </body>
table not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>

#data
<!DOCTYPE HTML><table><thead><tr><td></body>
#errors
unmatched </body>
table not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <tr>
|           <td>

#data
<!DOCTYPE HTML><isindex>
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <label>
|         "This is a searchable index. Insert your search keywords here: "
|         <input>
|           name="isindex"
|       <hr>

#data
<!DOCTYPE HTML><isindex/>
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <label>
|         "This is a searchable index. Insert your search keywords here: "
|         <input>
|           name="isindex"
|       <hr>

#data
<!DOCTYPE HTML><form><isindex/></form>
#errors
<isindex> in form
nestc
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>

#data
<!DOCTYPE HTML><isindex action="http://test/">
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       action="http://test/"
|       <hr>
|       <label>
|         "This is a searchable index. Insert your search keywords here: "
|         <input>
|           name="isindex"
|       <hr>

#data
<!DOCTYPE HTML><isindex prompt="Input:">
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <label>
|         "Input:"
|         <input>
|           name="isindex"
|       <hr>

#data
<!DOCTYPE HTML><isindex name="textarea">
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <label>
|         "This is a searchable index. Insert your search keywords here: "
|         <input>
|           name="isindex"
|       <hr>

#data
<!DOCTYPE HTML><isindex type="checkbox">
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <label>
|         "This is a searchable index. Insert your search keywords here: "
|         <input>
|           name="isindex"
|           type="checkbox"
|       <hr>

#data
<!DOCTYPE HTML><form><isindex></form>
#errors
isindex
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>

#data
<!DOCTYPE HTML><h1><a><h2>
#errors
unexpected end of file (no </h1></a></h2>)
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <h1>
|       <a>
|         <h2>

#data
<!DOCTYPE HTML><h1><p></h1>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <h1>
|       <p>

#data
<!DOCTYPE HTML><h1><li></h1>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <h1>
|       <li>

#data
<!DOCTYPE HTML><h1><dt></h1>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <h1>
|       <dt>

#data
<!DOCTYPE HTML><h1><dd></h1>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <h1>
|       <dd>

#data
<!DOCTYPE HTML><h1></h1>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <h1>

#data
  <p>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>

#data
  a
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     "a"

#data
<!---->  <p>
#errors
no DOCTYPE
#document
| <!--  -->
| <html>
|   <head>
|   <body>
|     <p>

#data
<!DOCTYPE HTML>  <p>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>

#data
<!DOCTYPE HTML><div></div></div>
#errors
unmatched </div>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <div>

#data
<!DOCTYPE HTML><div><span></div>
#errors
unclosed <span>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <div>
|       <span>

#data
<!DOCTYPE HTML><listing>
</listing>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <listing>

#data
<!DOCTYPE HTML><listing>
a</listing>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <listing>
|       "a"

#data
<!DOCTYPE HTML><listing>a
</listing>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <listing>
|       "a
"

#data
<!DOCTYPE HTML><listing>

</listing>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <listing>
|       "
"

#data
<!DOCTYPE HTML><listing> 
</listing>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <listing>
|       " 
"

#data
<!DOCTYPE HTML><pre><!---->
a</pre>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <pre>
|       <!--  -->
|       "
a"

#data
<!DOCTYPE HTML><listing><!---->
a</listing>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <listing>
|       <!--  -->
|       "
a"

#data
<!DOCTYPE HTML><textarea><!---->
a</textarea>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "<!---->
a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html> 
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     " "

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>  
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "  "

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>a
#errors
char after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html> a
#errors
char after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     " a"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>a 
#errors
char after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "a "

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>aaa
#errors
char after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "aaa"

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html><!---->
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
| <!--  -->

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html><!--aa-->
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
| <!-- aa -->

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html><!DOCTYPE HTML>
#errors
<!DOCTYPE> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html><p>
#errors
start tag after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <p>

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html></p>
#errors
end tag after </html>
</p> interpreted as <p></p>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     <p>

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>  <p>
#errors
start tag after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "  "
|     <p>

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body></html>aa<p>
#errors
char after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     "aa"
|     <p>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> 
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html>  
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   "  "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html>a
#errors
char after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> a
#errors
char after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html>a 
#errors
char after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><frameset></frameset></html>a</html>
#errors
char after </html>
</html> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><!---->
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
| <!--  -->

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><!--a-->
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
| <!-- a -->

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> <!---->
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "
| <!--  -->

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html>a<!---->
#errors
char after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
| <!--  -->

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html>a<!----> 
#errors
char after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "
| <!--  -->

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html>a<!----> <!--a-->
#errors
char after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "
| <!--  -->
| <!-- a -->

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><!DOCTYPE HTML>
#errors
<!DOCTYPE> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> <!DOCTYPE HTML>
#errors
<!DOCTYPE> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><p>
#errors
start tag after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> <p>
#errors
start tag after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html></p>
#errors
end tag after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> </p>
#errors
end tag after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html></html>
#errors
</html> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html>a</html>
#errors
char after </html>
</html> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><noframes></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   <noframes>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><noframes>
#errors
no </noframes>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   <noframes>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><noframes><p></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   <noframes>
|     "<p>"

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><noframes></noframes>a
#errors
char after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   <noframes>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html><noframes></noframes>a</html>
#errors
char after </html>
</html> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   <noframes>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset></frameset></html> <noframes></noframes> 
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|   " "
|   <noframes>
|   " "

#data
<!DOCTYPE HTML><frameset></frameset></html><frameset>
#errors
<frameset> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><frameset></frameset></html></frameset>
#errors
</frameset> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><frameset></frameset></html><frame>
#errors
<frame> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><frameset></frameset></html><xyz>
#errors
<xyz> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><frameset></frameset></html></xyz>
#errors
</xyz> after </html>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><html><head><title></title></head><frameset> </frameset></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <frameset>
|     " "

#data
<!DOCTYPE HTML><html><head><title></title></head><body></body> </html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   <body>
|     " "

#data
<!DOCTYPE HTML><frameset></frameset><noframes>
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>

#data
<!DOCTYPE HTML><frameset></frameset><noframes></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>

#data
<!DOCTYPE HTML><frameset></frameset><noframes>abcdefg
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "abcdefg"

#data
<!DOCTYPE HTML><frameset></frameset><noframes>abcdefg</noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "abcdefg"

#data
<!DOCTYPE HTML><frameset></frameset><noframes>aa<p>bb</noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "aa<p>bb"

#data
<!DOCTYPE HTML><frameset></frameset><noframes><noframes></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "<noframes>"

#data
<!DOCTYPE HTML><frameset></frameset><noframes></noframe></noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>
|   <noframes>
|     "</noframe>"

#data
<!DOCTYPE HTML><noframes>x</noframes><!---->
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noframes>
|       "x"
|     <!--  -->
|   <body>

#data
<!DOCTYPE HTML><head><noframes>x</noframes><!---->
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noframes>
|       "x"
|     <!--  -->
|   <body>

#data
<!DOCTYPE HTML><head></head><noframes>x</noframes><!---->
#errors
<noframes> after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noframes>
|       "x"
|   <!--  -->
|   <body>

#data
<!DOCTYPE HTML><body><noframes>x</noframes>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <noframes>
|       "x"

#data
</body></html><html a>
#errors
no <!DOCTYPE>
</body> before head
</html> before head
<html> in body
#document
| <html>
|   a=""
|   <head>
|   <body>

#data
</html>
#errors
no <!DOCTYPE>
</html> before head
#document
| <html>
|   <head>
|   <body>

#data
<!----></html>x
#errors
no <!DOCTYPE>
</html> before head
#document
| <!--  -->
| <html>
|   <head>
|   <body>
|     "x"

#data
<body></body></html><html a>
#errors
no <!DOCTYPE>
<html> after </html>
<html> is not the first start tag
#document
| <html>
|   a=""
|   <head>
|   <body>

#data
<frameset></frameset><html a></html>
#errors
no <!DOCTYPE>
<html> is not the first start tag
#document
| <html>
|   a=""
|   <head>
|   <frameset>

#data
<frameset></frameset></html><html a>
#errors
no <!DOCTYPE>
<html> after </html>
<html> is not the first start tag
#document
| <html>
|   a=""
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><html><head></head><frameset></frameset></html><html a>
#errors
<html> after </html>
<html> is not the first start tag
#document
| <!DOCTYPE HTML>
| <html>
|   a=""
|   <head>
|   <frameset>

#data
<frameset></frameset>
#errors
no <!DOCTYPE>
#document
| <html>
|   <head>
|   <frameset>

#data
<frameset></frameset></frameset>
#errors
no <!DOCTYPE>
unmatched </frameset>
#document
| <html>
|   <head>
|   <frameset>

#data
<frameset></frameset></frameset></frameset>
#errors
no <!DOCTYPE>
unmatched </frameset>
unmatched </frameset>
#document
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><table><caption>test</caption>a</table>
#errors
a in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <table>
|       <caption>
|         "test"

#data
<!DOCTYPE HTML><table><caption>test</table>a
#errors
no </caption>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "test"
|     "a"

#data
<!DOCTYPE HTML><table><caption>test</body>a
#errors
unmatched </body>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</col>a
#errors
unmatched </col>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</colgroup>a
#errors
unmatched </colgroup>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</html>a
#errors
unmatched </html>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</html> 
#errors
unmatched </html>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "test "

#data
<!DOCTYPE HTML><table><caption>test</tbody>a
#errors
unmatched </tbody>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</thead>a
#errors
unmatched </thead>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</tfoot>a
#errors
unmatched </tfoot>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</tr>a
#errors
unmatched </tr>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</td>a
#errors
unmatched </td>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>test</th>a
#errors
unmatched </th>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "testa"

#data
<!DOCTYPE HTML><table><caption>a<caption>b
#errors
no </caption>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "a"
|       <caption>
|         "b"

#data
<!DOCTYPE HTML><table><colgroup>  </table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>
|         "  "

#data
<!DOCTYPE HTML><table><colgroup></colgroup></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>

#data
<!DOCTYPE HTML><table><caption>a<col>b
#errors
no </caption>
"b" in table
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <colgroup>
|         <col>

#data
<!DOCTYPE HTML><table><caption>a<colgroup>b
#errors
no </caption>
"b" in colgroup
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <colgroup>

#data
<!DOCTYPE HTML><table><caption>a<thead>b
#errors
no </caption>
"b" in thead
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <thead>

#data
<!DOCTYPE HTML><table><caption>a<tbody>b
#errors
no </caption>
"b" in tbody
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <tbody>

#data
<!DOCTYPE HTML><table><caption>a<tfoot>b
#errors
no </caption>
"b" in tfoot
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <tfoot>

#data
<!DOCTYPE HTML><table><caption>a<tr>b
#errors
no </caption>
"b" in tr
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <caption>
|         "a"
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><caption>a<td>b
#errors
no </caption>
no <tr>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "a"
|       <tbody>
|         <tr>
|           <td>
|             "b"

#data
<!DOCTYPE HTML><table><caption>a<th>b
#errors
no </caption>
no <tr>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         "a"
|       <tbody>
|         <tr>
|           <th>
|             "b"

#data
<!DOCTYPE HTML>a<caption>b
#errors
<caption> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML><table><caption><span><tbody>
#errors
<caption> not closed
<span> not closed
<table> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <span>
|       <tbody>

#data
<!DOCTYPE HTML><table><caption><li><tbody>
#errors
<caption> not closed
<table> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <li>
|       <tbody>

#data
<!DOCTYPE HTML><table><caption><p><tbody>
#errors
<caption> not closed
<table> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <p>
|       <tbody>

#data
<!DOCTYPE HTML><table><caption><dt><tbody>
#errors
<caption> not closed
<table> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <dt>
|       <tbody>

#data
<!DOCTYPE HTML><table><caption><dd><tbody>
#errors
<caption> not closed
<table> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <dd>
|       <tbody>

#data
<!DOCTYPE HTML>a<col>b
#errors
<col> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<colgroup>b
#errors
<colgroup> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<frame>b
#errors
<frame> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<frameset>b
#errors
<frameset> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<frameset><frame>b
#errors
<frameset> in body
<frame> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<head>b
#errors
<head> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<head><link>b
#errors
<head> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <link>
|     "b"

#data
<!DOCTYPE HTML>a<option>b
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <option>
|       "b"

#data
<!DOCTYPE HTML>a<optgroup>b
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <optgroup>
|       "b"

#data
<!DOCTYPE HTML>a<tbody>b
#errors
<tbody> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<tbody><tr>b
#errors
<tbody> in body
<tr> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<thead>b
#errors
<thead> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<tfoot>b
#errors
<tfoot> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<tr>b
#errors
<tr> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<td>b
#errors
<td> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML>a<th>b
#errors
<th> in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "ab"

#data
<!DOCTYPE HTML><table><tr><td>a<tr><td>b</table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"
|         <tr>
|           <td>
|             "b"

#data
<!DOCTYPE HTML><table><tr><td>a<thead><tr><td>b</table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"
|       <thead>
|         <tr>
|           <td>
|             "b"

#data
<!DOCTYPE HTML><table><tr><td>a<tbody><tr><td>b</table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"
|       <tbody>
|         <tr>
|           <td>
|             "b"

#data
<!DOCTYPE HTML><table><tr><td>a<tfoot><tr><td>b</table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"
|       <tfoot>
|         <tr>
|           <td>
|             "b"

#data
<!DOCTYPE HTML><table><tr><td>a<caption>b</caption></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"
|       <caption>
|         "b"

#data
<!DOCTYPE HTML><table><tr><td>a</th></table>
#errors
unmatched </th>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"

#data
<!DOCTYPE HTML><table><tr><td>a</th>b</table>
#errors
unmatched </th>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</body>b
#errors
unmatched </body>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</html>b
#errors
unmatched </html>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</caption>b
#errors
unmatched </caption>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</colgroup>b
#errors
unmatched </colgroup>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</col>b
#errors
unmatched </col>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</thead>b
#errors
unmatched </thead>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</tfoot>b
#errors
unmatched </tfoot>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tfoot><tr><td>a</tbody>b
#errors
unmatched </tbody>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|         <tr>
|           <td>
|             "ab"

#data
<!DOCTYPE HTML><table><tr><td>a</tbody>b
#errors
"b" in table
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             "a"

#data
<!DOCTYPE HTML><table><thead><tr><td>a</thead>b
#errors
"b" in table
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <thead>
|         <tr>
|           <td>
|             "a"

#data
<!DOCTYPE HTML><table><tfoot><tr><td>a</tfoot>b
#errors
"b" in table
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "b"
|     <table>
|       <tfoot>
|         <tr>
|           <td>
|             "a"

#data
<!DOCTYPE HTML><table><tbody><caption>a</caption>
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|       <caption>
|         "a"

#data
<!DOCTYPE HTML><table><caption><li></caption></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <li>

#data
<!DOCTYPE HTML><table><caption><p></caption></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <p>

#data
<!DOCTYPE HTML><table><caption><dt></caption></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <dt>

#data
<!DOCTYPE HTML><table><caption><dd></caption></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <dd>

#data
<!DOCTYPE HTML><table><caption><p></table>
#errors
<caption> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <p>

#data
<!DOCTYPE HTML><table><caption><li></table>
#errors
<caption> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <li>

#data
<!DOCTYPE HTML><table><caption><dd></table>
#errors
<caption> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <dd>

#data
<!DOCTYPE HTML><table><caption><dt></table>
#errors
<caption> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <dt>

#data
<!DOCTYPE HTML><table><caption><span></table>
#errors
<caption> not closed
<span> not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <span>

#data
<!DOCTYPE HTML><table><tbody><thead><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|       <thead>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody><tfoot><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|       <tfoot>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody><tbody><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><thead><tfoot><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|       <tfoot>
|         <tr>

#data
<!DOCTYPE HTML><table><thead><tbody><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><thead><thead><tr></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|       <thead>
|         <tr>

#data
<!DOCTYPE HTML><table a><thead><table b>
#errors
<table> in table row group
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|       <thead>
|     <table>
|       b=""

#data
<!DOCTYPE HTML><table a><tbody><table b>
#errors
<table> in table row group
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|       <tbody>
|     <table>
|       b=""

#data
<!DOCTYPE HTML><table a><tfoot><table b>
#errors
<table> in table row group
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|       <tfoot>
|     <table>
|       b=""

#data
<!DOCTYPE HTML><table a><tr><table b>
#errors
<table> in table row
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|       <tbody>
|         <tr>
|     <table>
|       b=""

#data
<!DOCTYPE HTML><table><thead></tbody><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <tr>

#data
<!DOCTYPE HTML><table><thead></thead><tr>
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><thead></tfoot><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></thead><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></tbody><tr>
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></tfoot><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tfoot></thead><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|         <tr>

#data
<!DOCTYPE HTML><table><tfoot></tbody><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|         <tr>

#data
<!DOCTYPE HTML><table><tfoot></tfoot><tr>
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></tr><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></body><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></html><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></td><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></th><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></caption><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></col><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tbody></colgroup><tr>
#errors
unmatched end tag
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tr><form><td></td></form></table>
#errors
<form> in table
form not closed
</form> in table
ubmatched </form>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<!DOCTYPE HTML><table><tr><td></th></table>
#errors
unmatched </th>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<!DOCTYPE HTML><table><tr><th></td></table>
#errors
unmatched </td>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>

#data
<!DOCTYPE HTML><table><tr><td><dd></td></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <dd>

#data
<!DOCTYPE HTML><table><tr><td><dt></td></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <dt>

#data
<!DOCTYPE HTML><table><tr><td><li></td></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <li>

#data
<!DOCTYPE HTML><table><tr><td><p></td></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <p>

#data
<!DOCTYPE HTML><table><tr><th><p></th></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             <p>

#data
<!DOCTYPE HTML><body></h1>
#errors
unmatched </h1>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<p><font><form> </form>
#errors
1,4,no DOCTYPE
1,16,not closed:font
unexpected end-of-file
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <font>
|     <form>
|       <font>
|         " "

#data
<p><font><form> </form>a
#errors
no DOCTYPE
<font> not closed before <form>
<font> not closed before </form>
unexpected end-of-file (<font> not closed before end-of-file)
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <font>
|     <form>
|       <font>
|         " "
|     <font>
|       "a"

#data
<table a><tr><table b>
#errors
1,10,no DOCTYPE
1,22,not closed:tr
1,22,not closed:table
#document
| <html>
|   <head>
|   <body>
|     <table>
|       a=""
|       <tbody>
|         <tr>
|     <table>
|       b=""

#data
<!DOCTYPE HTML>foo<html x>
#errors
<html> in body
#document
| <!DOCTYPE HTML>
| <html>
|   x=""
|   <head>
|   <body>
|     "foo"

#data
<!DOCTYPE HTML><select><option><option>
#errors
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|       <option>

#data
<!DOCTYPE HTML><select></p>
#errors
</p> in <select>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE HTML><select></br>
#errors
</br> in <select>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE HTML><select></base>
#errors
</base> in <select>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE HTML><select></div>
#errors
</div> in <select>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE HTML><select></option>
#errors
unmatched </option>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE HTML><select><option></option></select>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>

#data
<!DOCTYPE HTML><select><optgroup></option></select>
#errors
</option> in <optgroup>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <optgroup>

#data
<!DOCTYPE HTML><select></table>aa
#errors
unmatched </table>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "aa"

#data
<!DOCTYPE HTML><select></tbody>aa
#errors
unmatched </tbody>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "aa"

#data
<!DOCTYPE HTML><select></thead>aa
#errors
unmatched </thead>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "aa"

#data
<!DOCTYPE HTML><select></tfoot>aa
#errors
unmatched </tfoot>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "aa"

#data
<!DOCTYPE HTML><select></tr>aa
#errors
unmatched </tr>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "aa"

#data
<!DOCTYPE HTML><select></td>aa
#errors
unmatched </td>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "aa"

#data
<!DOCTYPE HTML><select></th>aa
#errors
unmatched </th>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "aa"

#data
<!DOCTYPE HTML><select></caption>aa
#errors
unmatched </caption>
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "aa"

#data
<!DOCTYPE HTML><table><tr><td><select></tbody>a
#errors
no </select>
char in <table>
no </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>

#data
<!DOCTYPE HTML><table><tr><td><select></table>a
#errors
no </select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|     "a"

#data
<!DOCTYPE HTML><table><tr><td><select></thead>a
#errors
unmatched </thead>
unexpected end of file (no </select>, no </table>)
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               "a"

#data
<!DOCTYPE HTML><table><tr><td><select></tfoot>a
#errors
unmatched </tfoot>
unexpected end of file (no </select>, no </table>)
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               "a"

#data
<!DOCTYPE HTML><table><tfoot><tr><td><select></thead>a
#errors
unmatched </thead>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tfoot>
|         <tr>
|           <td>
|             <select>
|               "a"

#data
<!DOCTYPE HTML><table><tr><td><select></tr>a
#errors
no </select>
char in <tbody>
no </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>

#data
<!DOCTYPE HTML><table><tr><td><select></td>a
#errors
no </select>
char in <tr>
no </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>

#data
<!DOCTYPE HTML><table><tr><td><select></th>a
#errors
unmatched </th>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               "a"

#data
<!DOCTYPE HTML><table><select></table>a
#errors
<select> in <table>
no </select> before </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|     <table>
|     "a"

#data
<!DOCTYPE HTML><table><select><option>b</table>a
#errors
<select> in <table>
no </select> before </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         "b"
|     <table>
|     "a"

#data
<!DOCTYPE HTML><table><tfoot><select></table>a
#errors
<select> in <tfoot>
no </select> before </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|     <table>
|       <tfoot>
|     "a"

#data
<!DOCTYPE HTML><table><select><option>b</tfoot>a
#errors
<select> in <table>
unmatched </tfoot>
no </table> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         "ba"
|     <table>

#data
<!DOCTYPE HTML><table><tr><select></table>a
#errors
<select> in <tr>
no </select> before </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|     <table>
|       <tbody>
|         <tr>
|     "a"

#data
<!DOCTYPE HTML><table><tr><select></tr>a</table>
#errors
<select> in <tr>
no </select> before </tr>
char in <tbody>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|     "a"
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><table><tr><td><select><option>b</table>a
#errors
no </select> before </table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|               <option>
|                 "b"
|     "a"

#data
<!DOCTYPE HTML><select><input>x
#errors
no </select> before </input>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|     <input>
|     "x"

#data
<!DOCTYPE HTML><select><option>a<input>b
#errors
no </select> before <input>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         "a"
|     <input>
|     "b"

#data
<!DOCTYPE HTML><table><select><input></table>a
#errors
<select> in <table>
no </select> before <input>
<input> in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|     <input>
|     <table>
|     "a"

#data
<!DOCTYPE HTML><table><select><option>b<input></table>a
#errors
<select> in <table>
no </select> before <input>
<input> in <table>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         "b"
|     <input>
|     <table>
|     "a"

#data
<!DOCTYPE HTML><select><textarea></textarea>
#errors
<textarea> in <select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|     <textarea>

#data
<!DOCTYPE HTML><select><option><textarea></textarea>
#errors
<textarea> in <select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|     <textarea>

#data
<!DOCTYPE HTML><select><optgroup><textarea></textarea>
#errors
<textarea> in <select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <optgroup>
|     <textarea>

#data
<!DOCTYPE HTML><table><tr><td><select><textarea></textarea></table>
#errors
<textarea> in <select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|             <textarea>

#data
<!DOCTYPE HTML><select><optgroup><keygen></keygen>
#errors
<keygen> in <select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <optgroup>
|     <keygen>

#data
<!DOCTYPE HTML><table><tr><td><select><keygen></keygen></table>
#errors
<keygen> in <select>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|             <keygen>

#data
<!DOCTYPE HTML><select><table></table>
#errors
<table> in <select>
</table> in <select>
no </select> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE HTML><select><option><table></table>x
#errors
<table> in <select>
</table> in <select>
no </select> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         "x"

#data
<!DOCTYPE HTML><table><tr><td><select><table></table>x
#errors
no </select> before <table>
no </table> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|             <table>
|             "x"

#data
<!DOCTYPE HTML><select><option><caption>x
#errors
<caption> in <select>
no </select> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>
|         "x"

#data
<!DOCTYPE HTML><table><tr><td><select><caption>x
#errors
no </select> before <caption>
no </caption></table> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|       <caption>
|         "x"

#data
<!DOCTYPE HTML><select><tbody>x
#errors
<tbody> in <select>
no </select> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "x"

#data
<!DOCTYPE HTML><table><tr><td><select><tbody><tr><td>x
#errors
no </select> before <tbody>
no </table> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|       <tbody>
|         <tr>
|           <td>
|             "x"

#data
<!DOCTYPE HTML><select><tr>x
#errors
<tr> in <select>
no </select> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "x"

#data
<!DOCTYPE HTML><table><tr><td><select><tr><td>x
#errors
no </select> before <tr>
no </table> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|         <tr>
|           <td>
|             "x"

#data
<!DOCTYPE HTML><select><td>x
#errors
<td> in <select>
no </select> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       "x"

#data
<!DOCTYPE HTML><table><tr><td><select><td>x
#errors
no </select> before <table>
no </table> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <select>
|           <td>
|             "x"

#data
<!DOCTYPE HTML><p><plaintext>
#errors
unexpected end of file (no </plaintext>)
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|     <plaintext>

#data
<!DOCTYPE HTML><p><h1>
#errors
unexpected end of file (no </h1>)
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|     <h1>

#data
<!DOCTYPE HTML><p><table></table>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|     <table>

#data
<!DOCTYPE HTML><textarea></textarea>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <textarea>

#data
<!DOCTYPE HTML><textarea>
#errors
no </textarea>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <textarea>

#data
<!DOCTYPE HTML><span></html>
#errors
no </span>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <span>

#data
<!DOCTYPE HTML><unknown><p></unknown>
#errors
unmatched </unknown>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <unknown>
|       <p>

#data
<!DOCTYPE HTML><unknown><p></unknown>xx
#errors
unmatched </unknown>
unexpected end of file (no </unknown>)
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <unknown>
|       <p>
|         "xx"

#data
<!DOCTYPE HTML><p></unknown>xx
#errors
unmatched </unknown>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       "xx"

#data
<!DOCTYPE HTML><b>a<i>b<applet>c</b>d</i>e
#errors
1,37,unmatched end tag:b
1,42,unmatched end tag:i
no </applet> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <b>
|       "a"
|       <i>
|         "b"
|         <applet>
|           "cde"

#data
<!DOCTYPE HTML><li>a<applet>b</li>c
#errors
1,35,unmatched end tag:li
no </applet> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <li>
|       "a"
|       <applet>
|         "bc"

#data
<!DOCTYPE HTML><dt>a<applet>b</dt>c
#errors
1,35,unmatched end tag:dt
no </applet> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dt>
|       "a"
|       <applet>
|         "bc"

#data
<!DOCTYPE HTML><dd>a<applet>b</dd>c
#errors
1,35,unmatched end tag:dd
no </applet> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dd>
|       "a"
|       <applet>
|         "bc"

#data
<!DOCTYPE HTML><x>a<applet>b</x>c
#errors
1,33,unmatched end tag:x
no </applet> at the end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <x>
|       "a"
|       <applet>
|         "bc"

#data
<!DOCTYPE HTML><p><applet><address>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <address>

#data
<!DOCTYPE HTML><p><applet><h1>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <h1>

#data
<!DOCTYPE HTML><p><applet><pre>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <pre>

#data
<!DOCTYPE HTML><p><applet><listing>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <listing>

#data
<!DOCTYPE HTML><p><applet><div>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <div>

#data
<!DOCTYPE HTML><p><applet><form>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <form>

#data
<!DOCTYPE HTML><p><applet><li>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <li>

#data
<!DOCTYPE HTML><p><applet><dt>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <dt>

#data
<!DOCTYPE HTML><p><applet><dd>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <dd>

#data
<!DOCTYPE HTML><p><applet><plaintext>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <plaintext>

#data
<!DOCTYPE HTML><p><applet><table>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <table>

#data
<!DOCTYPE HTML><p><applet><hr>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <hr>

#data
<!DOCTYPE HTML><nobr><applet><nobr>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <nobr>
|       <applet>
|         <nobr>

#data
<!DOCTYPE HTML><button><applet><button>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <button>
|       <applet>
|         <button>

#data
<!DOCTYPE HTML><address><applet></address>x
#errors
unmatched end tag
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <address>
|       <applet>
|         "x"

#data
<!DOCTYPE HTML><h1><applet></h1>x
#errors
unmatched end tag
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <h1>
|       <applet>
|         "x"

#data
<!DOCTYPE HTML><button><applet></button>x
#errors
unmatched end tag
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <button>
|       <applet>
|         "x"

#data
<!DOCTYPE HTML><object><applet></object>x
#errors
unmatched end tag
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <object>
|       <applet>
|         "x"

#data
<!DOCTYPE HTML><p><applet></p>x
#errors
unmatched end tag </p>
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <applet>
|         <p>
|         "x"

#data
<!DOCTYPE HTML><dt><applet></dt>x
#errors
unmatched end tag
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dt>
|       <applet>
|         "x"

#data
<!DOCTYPE HTML><dd><applet></dd>x
#errors
unmatched end tag
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dd>
|       <applet>
|         "x"

#data
<!DOCTYPE HTML><pre><applet></pre>x
#errors
unmatched end tag
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <pre>
|       <applet>
|         "x"

#data
<!DOCTYPE HTML><form><applet></form>x
#errors
unmatched end tag
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <form>
|       <applet>
|         "x"

#data
<!DOCTYPE HTML><applet><applet></applet>x
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <applet>
|       <applet>
|       "x"

#data
#errors
not HTML5
#document
| <html>
|   <head>
|   <body>

#data
<!---->
#errors
not HTML5
#document
| <!--  -->
| <html>
|   <head>
|   <body>

#data
    
#errors
not HTML5
#document
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><!---->
#errors
#document
| <!DOCTYPE HTML>
| <!--  -->
| <html>
|   <head>
|   <body>

#data
<html>
#errors
not HTML5
#document
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
a
#errors
not HTML5
#document
| <html>
|   <head>
|   <body>
|     "a"

#data
<!DOCTYPE HTML><html><head>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html><head><noscript>
#errors
no </noscript>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>

#data
<!DOCTYPE HTML><html><head></head>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html><head></head><body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html><head></head><body><table>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>

#data
<!DOCTYPE HTML><html><head></head><body><table><span>
#errors
<span> in <table>
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <span>
|     <table>

#data
<!DOCTYPE HTML><html><head></head><body><table><caption>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>

#data
<!DOCTYPE HTML><html><head></head><body><table><colgroup>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>

#data
<!DOCTYPE HTML><html><head></head><body><table><thead>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <thead>

#data
<!DOCTYPE HTML><html><head></head><body><table><tr>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML><html><head></head><body><table><tr><td>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<!DOCTYPE HTML><html><head></head><body><select>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>

#data
<!DOCTYPE HTML><html><head></head><body><select><option>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <option>

#data
<!DOCTYPE HTML><html><head></head><body><select><optgroup>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <select>
|       <optgroup>

#data
<!DOCTYPE HTML><html><head></head><body></body>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html><head></head><body></body></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML><html><head></head><frameset>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><html><head></head><frameset></frameset>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><html><head></head><frameset></frameset></html>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><html><head></head><body><li>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <li>

#data
<!DOCTYPE HTML><html><head></head><body><dt>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dt>

#data
<!DOCTYPE HTML><html><head></head><body><dd>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dd>

#data
<!DOCTYPE HTML><html><head></head><body><p>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>

#data
<!DOCTYPE HTML><html><head></head><body><span>
#errors
unexpected end-of-file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <span>

#data
<!DOCTYPE HTML><ruby><ol><li><p>a<rt>b
#errors
no </ol>
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <ol>
|         <li>
|           <p>
|             "a"
|       <rt>
|         "b"

#data
<!DOCTYPE HTML><ruby><marquee><p>a<rt>b
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <marquee>
|         <p>
|           "a"
|           <rt>
|             "b"

#data
<!DOCTYPE HTML><ruby>a<rt>b<p>c
#errors
unexpected end of file
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ruby>
|       "a"
|       <rt>
|         "b"
|         <p>
|           "c"

#data
<!DOCTYPE HTML><p><ruby>a<rt>b<p>c
#errors
no </ruby>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <p>
|       <ruby>
|         "a"
|         <rt>
|           "b"
|     <p>
|       "c"

#data
<!DOCTYPE HTML><ruby>a<rt>b</ruby>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ruby>
|       "a"
|       <rt>
|         "b"

#data
<!DOCTYPE HTML><ruby><rb>a</rb><rt>b</rt></ruby>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <rb>
|         "a"
|       <rt>
|         "b"

#data
<!DOCTYPE HTML><ruby><rb>a<rt>b</ruby>
#errors
no </rb>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <rb>
|         "a"
|       <rt>
|         "b"

#data
<!DOCTYPE HTML><ruby>a<rp>b<rt>c<rp>d</ruby>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ruby>
|       "a"
|       <rp>
|         "b"
|       <rt>
|         "c"
|       <rp>
|         "d"

#data
<!DOCTYPE HTML><ruby>a<rt>b</rt></ruby>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ruby>
|       "a"
|       <rt>
|         "b"

#data
<!DOCTYPE HTML><ruby>a<rp>b</rp><rt>c</rt><rp>d</rp></ruby>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ruby>
|       "a"
|       <rp>
|         "b"
|       <rt>
|         "c"
|       <rp>
|         "d"

#data
<!DOCTYPE HTML><ruby><rb>a</rb><rp>b</rp><rt>c</rt><rp>d</rp></ruby>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <rb>
|         "a"
|       <rp>
|         "b"
|       <rt>
|         "c"
|       <rp>
|         "d"

#data
<!DOCTYPE HTML><ruby><rbc><rb>a</rb><rb>b</rb></rbc><rtc><rt>c</rt><rt>d</rt></rtc><rtc><rt>e</rt><rt>f</rt></rtc></ruby>
#errors
no </rtc>
unexpected </rtc>
no </rtc>
unexpected </rtc>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ruby>
|       <rbc>
|         <rb>
|           "a"
|         <rb>
|           "b"
|       <rtc>
|       <rt>
|         "c"
|       <rt>
|         "d"
|       <rtc>
|       <rt>
|         "e"
|       <rt>
|         "f"

#data
<!DOCTYPE HTML>
... <ruby>
 SAI <rt> sai </rt>
 TOU <rt> tou </rt>
 NOBU <rt> nobu </rt>
 O <rt> o         </rt>
</ruby> ...
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "... "
|     <ruby>
|       "
 SAI "
|       <rt>
|         " sai "
|       "
 TOU "
|       <rt>
|         " tou "
|       "
 NOBU "
|       <rt>
|         " nobu "
|       "
 O "
|       <rt>
|         " o         "
|       "
"
|     " ..."

#data
<!DOCTYPE HTML><div><rt></div>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <div>
|       <rt>

#data
<!DOCTYPE HTML><div><rp></div>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <div>
|       <rp>

#data
<!DOCTYPE HTML><div><rt><rp></div>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <div>
|       <rt>
|         <rp>

#data
<!DOCTYPE HTML><ul><li><rt></ul>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <rt>

#data
<!DOCTYPE HTML><svg>
#errors
end of file in foreign content
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <svg svg>

#data
<!DOCTYPE HTML><svg><g>
#errors
end of file in foreign content
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg g>

#data
<!DOCTYPE HTML><svg><switch requiredfeatures></switch></svg>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg switch>
|         requiredFeatures=""

#data
<!DOCTYPE HTML><svg><rect x="1" y="1" width="2" height="2" fill="black"><animate attributeNAME="x" begin="0s" dur="9s" fill="freeze" from="300" to="0"/></rect></svg>
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <svg svg>
|       <svg rect>
|         fill="black"
|         height="2"
|         width="2"
|         x="1"
|         y="1"
|         <svg animate>
|           attributeName="x"
|           begin="0s"
|           dur="9s"
|           fill="freeze"
|           from="300"
|           to="0"

#data
<!DOCTYPE HTML><svg><dt>
#errors
<dt> in <svg>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <svg svg>
|     <dt>

#data
<!DOCTYPE HTML><svg><h4>
#errors
<h4> in <svg>
h4 not closed
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <svg svg>
|     <h4>

#data
<!DOCTYPE HTML><svg><listing>
#errors
<listing> in <svg>
no </listing>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <svg svg>
|     <listing>

#data
<!DOCTYPE HTML><svg><ol></ol>
#errors
<ol> in <svg>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <svg svg>
|     <ol>

#data
<!DOCTYPE HTML><svg><strike></strike>
#errors
<strike> in <svg>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <svg svg>
|     <strike>

#data escaped
\u000b<p>
#errors
no DOCTYPE
U+000B
#document escaped
| <html>
|   <head>
|   <body>
|     "\u000b"
|     <p>

#data escaped
\u000babc
#errors
no DOCTYPE
U+000B
#document escaped
| <html>
|   <head>
|   <body>
|     "\u000babc"

#data escaped
<!DOCTYPE HTML>\u000b<p>
#errors
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "\u000b"
|     <p>

#data escaped
<!DOCTYPE HTML>\u000babc
#errors
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "\u000babc"

#data escaped
<!DOCTYPE HTML><html>\u000b<p>
#errors
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "\u000b"
|     <p>

#data escaped
<!DOCTYPE HTML><html>\u000babc
#errors
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "\u000babc"

#data escaped
<!DOCTYPE HTML><head>\u000b<p>
#errors
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "\u000b"
|     <p>

#data escaped
<!DOCTYPE HTML><head>\u000babc
#errors
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "\u000babc"

#data escaped
<!DOCTYPE HTML><head></head>\u000b<p>
#errors
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "\u000b"
|     <p>

#data escaped
<!DOCTYPE HTML><head></head>\u000babc
#errors
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "\u000babc"

#data escaped
<!DOCTYPE HTML><head><noscript>\u000b<p>
#errors
no </noscript>
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>
|     "\u000b"
|     <p>

#data escaped
<!DOCTYPE HTML><head><noscript>\u000babc
#errors
no </noscript>
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noscript>
|   <body>
|     "\u000babc"

#data escaped
<!DOCTYPE HTML><table>\u000b<p></table>
#errors
char in table
start tag in table
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "\u000b"
|     <p>
|     <table>

#data escaped
<!DOCTYPE HTML><table>\u000babc</table>
#errors
char in table
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "\u000babc"
|     <table>

#data escaped
<!DOCTYPE HTML><table><colgroup>\u000b<p></table>
#errors
char in table
start tag in table
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "\u000b"
|     <p>
|     <table>
|       <colgroup>

#data escaped
<!DOCTYPE HTML><table><colgroup>\u000babc</table>
#errors
char in table
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "\u000babc"
|     <table>
|       <colgroup>

#data escaped
<!DOCTYPE HTML><body></body>\u000b<p>
#errors
char after body
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "\u000b"
|     <p>

#data escaped
<!DOCTYPE HTML><body></body>\u000babc
#errors
char after body
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "\u000babc"

#data escaped
<!DOCTYPE HTML><frameset>\u000b<p></frameset>
#errors
char in frameset
start tag in frameset
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data escaped
<!DOCTYPE HTML><frameset>\u000babc</frameset>
#errors
char in frameset
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data escaped
<!DOCTYPE HTML><frameset></frameset>\u000b<p>
#errors
char after frameset
tag after frameset
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data escaped
<!DOCTYPE HTML><frameset>\u000babc</frameset>
#errors
char after frameset
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data escaped
<!DOCTYPE HTML><body></body></html>\u000b<p>
#errors
char after body html
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "\u000b"
|     <p>

#data escaped
<!DOCTYPE HTML><body></body></html>\u000babc
#errors
char after body html
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "\u000babc"

#data escaped
<!DOCTYPE HTML><frameset></frameset></html>\u000b<p>
#errors
char after frameset html
char after frameset html
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data escaped
<!DOCTYPE HTML><frameset></frameset></html>\u000babc
#errors
char after frameset html
U+000B
#document escaped
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <frameset>

#data
<!DOCTYPE HTML><ul><li><address><li>
#errors
no </address>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <address>
|       <li>

#data
<!DOCTYPE HTML><ul><li><address><p><li>
#errors
no </address>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <address>
|           <p>
|       <li>

#data
<!DOCTYPE HTML><ul><li><div><li>
#errors
no </div>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <div>
|       <li>

#data
<!DOCTYPE HTML><ul><li><div><p><li>
#errors
no </div>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <div>
|           <p>
|       <li>

#data
<!DOCTYPE HTML><ul><li><p><li>
#errors
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <p>
|       <li>

#data
<!DOCTYPE HTML><ul><li><span><li>
#errors
no </span>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <span>
|       <li>

#data
<!DOCTYPE HTML><ul><li><mark><li>
#errors
no </mark>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <mark>
|       <li>

#data
<!DOCTYPE HTML><ul><li><time><li>
#errors
no </time>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <time>
|       <li>

#data
<!DOCTYPE HTML><ul><li><option><li>
#errors
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <option>
|       <li>

#data
<!DOCTYPE HTML><ul><li><optgroup><li>
#errors
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <optgroup>
|       <li>

#data
<!DOCTYPE HTML><ul><li><rp><li>
#errors
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <rp>
|       <li>

#data
<!DOCTYPE HTML><ul><li><rt><li>
#errors
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <rt>
|       <li>

#data
<!DOCTYPE HTML><ul><li><unknown><li>
#errors
no </unknown>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <unknown>
|       <li>

#data
<!DOCTYPE HTML><ul><li><a><li>
#errors
no </a>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <a>
|       <li>

#data
<!DOCTYPE HTML><ul><li><b><li>
#errors
no </b>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <b>
|       <li>

#data
<!DOCTYPE HTML><ul><li><big><li>
#errors
no </big>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <big>
|       <li>

#data
<!DOCTYPE HTML><ul><li><em><li>
#errors
no </em>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <em>
|       <li>

#data
<!DOCTYPE HTML><ul><li><font><li>
#errors
no </font>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <font>
|       <li>

#data
<!DOCTYPE HTML><ul><li><i><li>
#errors
no </i>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <i>
|       <li>

#data
<!DOCTYPE HTML><ul><li><nobr><li>
#errors
no </nobr>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <nobr>
|       <li>

#data
<!DOCTYPE HTML><ul><li><s><li>
#errors
no </s>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <s>
|       <li>

#data
<!DOCTYPE HTML><ul><li><small><li>
#errors
no </small>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <small>
|       <li>

#data
<!DOCTYPE HTML><ul><li><strike><li>
#errors
no </strike>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <strike>
|       <li>

#data
<!DOCTYPE HTML><ul><li><strong><li>
#errors
no </strong>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <strong>
|       <li>

#data
<!DOCTYPE HTML><ul><li><tt><li>
#errors
no </tt>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <tt>
|       <li>

#data
<!DOCTYPE HTML><ul><li><u><li>
#errors
no </u>
no </ul>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <u>
|       <li>

#data
<!DOCTYPE HTML><ul><li><blockquote><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <blockquote>
|           <li>

#data
<!DOCTYPE HTML><ul><li><center><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <center>
|           <li>

#data
<!DOCTYPE HTML><ul><li><dd><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <dd>
|           <li>

#data
<!DOCTYPE HTML><ul><li><dir><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <dir>
|           <li>

#data
<!DOCTYPE HTML><ul><li><dl><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <dl>
|           <li>

#data
<!DOCTYPE HTML><ul><li><dt><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <dt>
|           <li>

#data
<!DOCTYPE HTML><ul><li><fieldset><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <fieldset>
|           <li>

#data
<!DOCTYPE HTML><ul><li><form><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <form>
|           <li>

#data
<!DOCTYPE HTML><ul><li><h1><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <h1>
|           <li>

#data
<!DOCTYPE HTML><ul><li><menu><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <menu>
|           <li>

#data
<!-- NOTE: Scripting disabled --><!DOCTYPE HTML><ul><li><noscript><li>
#errors
eof in body
#document
| <!--  NOTE: Scripting disabled  -->
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <noscript>
|           <li>

#data
<!DOCTYPE HTML><ul><li><ol><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <ol>
|           <li>

#data
<!DOCTYPE HTML><ul><li><pre><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <pre>
|           <li>

#data
<!DOCTYPE HTML><ul><li><listing><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <listing>
|           <li>

#data
<!DOCTYPE HTML><ul><li><ul><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <ul>
|           <li>

#data
<!DOCTYPE HTML><ul><li><article><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <article>
|           <li>

#data
<!DOCTYPE HTML><ul><li><aside><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <aside>
|           <li>

#data
<!DOCTYPE HTML><ul><li><datagrid><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <datagrid>
|           <li>

#data
<!DOCTYPE HTML><ul><li><details><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <details>
|           <li>

#data
<!DOCTYPE HTML><ul><li><dialog><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <dialog>
|           <li>

#data
<!DOCTYPE HTML><ul><li><figure><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <figure>
|           <li>

#data
<!DOCTYPE HTML><ul><li><footer><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <footer>
|           <li>

#data
<!DOCTYPE HTML><ul><li><header><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <header>
|           <li>

#data
<!DOCTYPE HTML><ul><li><nav><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <nav>
|           <li>

#data
<!DOCTYPE HTML><ul><li><section><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <section>
|           <li>

#data
<!DOCTYPE HTML><ul><li><applet><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <applet>
|           <li>

#data
<!DOCTYPE HTML><ul><li><button><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <button>
|           <li>

#data
<!DOCTYPE HTML><ul><li><marquee><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <marquee>
|           <li>

#data
<!DOCTYPE HTML><ul><li><object><li>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <object>
|           <li>

#data
<!DOCTYPE HTML><ul><li><table><li>
#errors
<li> in <table>
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <li>
|         <table>

#data
<!DOCTYPE HTML><ul><li><svg><foreignobject><li>
#errors
EOF in foreign content
unexpected EOF
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <svg svg>
|           <svg foreignObject>
|             <li>

#data
<!DOCTYPE HTML><dl><dt><address><dt>
#errors
no </address>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <address>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><address><p><dt>
#errors
no </address>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <address>
|           <p>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><div><dt>
#errors
no </div>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <div>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><div><p><dt>
#errors
no </div>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <div>
|           <p>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><p><dt>
#errors
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <p>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><span><dt>
#errors
no </span>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <span>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><mark><dt>
#errors
no </mark>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <mark>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><time><dt>
#errors
no </time>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <time>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><option><dt>
#errors
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <option>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><optgroup><dt>
#errors
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <optgroup>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><rt><dt>
#errors
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <rt>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><rp><dt>
#errors
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <rp>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><unknown><dt>
#errors
no </unknown>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <unknown>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><a><dt>
#errors
no </a>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <a>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><b><dt>
#errors
no </b>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <b>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><big><dt>
#errors
no </big>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <big>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><em><dt>
#errors
no </em>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <em>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><font><dt>
#errors
no </font>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <font>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><i><dt>
#errors
no </i>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <i>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><nobr><dt>
#errors
no </nobr>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <nobr>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><s><dt>
#errors
no </s>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <s>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><small><dt>
#errors
no </small>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <small>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><strike><dt>
#errors
no </strike>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <strike>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><strong><dt>
#errors
no </strong>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <strong>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><tt><dt>
#errors
no </tt>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <tt>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><u><dt>
#errors
no </u>
no </dl>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <u>
|       <dt>

#data
<!DOCTYPE HTML><dl><dt><blockquote><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <blockquote>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><center><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <center>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><dir><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <dir>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><dl><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <dl>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><li><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <li>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><fieldset><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <fieldset>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><form><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <form>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><h1><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <h1>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><menu><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <menu>
|           <dt>

#data
<!-- NOTE: Scripting disabled --><!DOCTYPE HTML><dl><dt><noscript><dt>
#errors
eof in body
#document
| <!--  NOTE: Scripting disabled  -->
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <noscript>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><ol><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <ol>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><pre><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <pre>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><listing><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <listing>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><dl><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <dl>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><article><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <article>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><aside><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <aside>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><datagrid><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <datagrid>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><details><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <details>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><dialog><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <dialog>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><figure><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <figure>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><footer><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <footer>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><header><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <header>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><nav><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <nav>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><section><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <section>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><applet><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <applet>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><button><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <button>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><marquee><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <marquee>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><object><dt>
#errors
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <object>
|           <dt>

#data
<!DOCTYPE HTML><dl><dt><table><dt>
#errors
<dt> in <table>
eof in body
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dt>
|         <dt>
|         <table>

#data
<!DOCTYPE HTML><dl><dd><svg><foreignobject><dd>
#errors
EOF in foreign content
unexpected EOF
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         <svg svg>
|           <svg foreignObject>
|             <dd>

#data
<span><address></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <address>
|         "a"

#data
<span><article></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <article>
|         "a"

#data
<span><aside></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <aside>
|         "a"

#data
<span><blockquote></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <blockquote>
|         "a"

#data
<span><center></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <center>
|         "a"

#data
<span><datagrid></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <datagrid>
|         "a"

#data
<span><details></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <details>
|         "a"

#data
<span><dialog></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <dialog>
|         "a"

#data
<span><dir></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <dir>
|         "a"

#data
<span><div></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <div>
|         "a"

#data
<span><dl></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <dl>
|         "a"

#data
<span><dd></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <dd>
|         "a"

#data
<span><dt></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <dt>
|         "a"

#data
<span><fieldset></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <fieldset>
|         "a"

#data
<span><figure></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <figure>
|         "a"

#data
<span><footer></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <footer>
|         "a"

#data
<span><form></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <form>
|         "a"

#data
<span><h2></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <h2>
|         "a"

#data
<span><header></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <header>
|         "a"

#data
<span><li></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <li>
|         "a"

#data
<span><menu></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <menu>
|         "a"

#data
<span><noscript></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <noscript>
|         "a"

#data
<span><ol></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <ol>
|         "a"

#data
<span><p></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <p>
|         "a"

#data
<span><pre></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <pre>
|         "a"

#data
<span><listing></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <listing>
|         "a"

#data
<span><section></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <section>
|         "a"

#data
<span><ul></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <ul>
|         "a"

#data
<span><applet></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <applet>
|         "a"

#data
<span><button></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <button>
|         "a"

#data
<span><marquee></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <marquee>
|         "a"

#data
<span><object></span>a
#errors
no DOCTYPE
unmatched end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <object>
|         "a"

#data
<span><svg><foreignobject></span>a
#errors
no DOCTYPE
unmatched end tag
EOF in foreign content
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <svg svg>
|         <svg foreignObject>
|           "a"

#data
<span><a></span>a
#errors
no DOCTYPE
element not closed before ancestor's end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <a>
|     <a>
|       "a"

#data
<span><b></span>a
#errors
no DOCTYPE
element not closed before ancestor's end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <b>
|     <b>
|       "a"

#data
<span><big></span>a
#errors
no DOCTYPE
element not closed before ancestor's end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <big>
|     <big>
|       "a"

#data
<span><em></span>a
#errors
no DOCTYPE
element not closed before ancestor's end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <em>
|     <em>
|       "a"

#data
<span><font></span>a
#errors
no DOCTYPE
element not closed before ancestor's end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <font>
|     <font>
|       "a"

#data
<span><i></span>a
#errors
no DOCTYPE
element not closed before ancestor's end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <i>
|     <i>
|       "a"

#data
<span><nobr></span>a
#errors
no DOCTYPE
element not closed before ancestor's end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <nobr>
|     <nobr>
|       "a"

#data
<span><s></span>a
#errors
no DOCTYPE
element not closed before ancestor's end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <s>
|     <s>
|       "a"

#data
<span><small></span>a
#errors
no DOCTYPE
element not closed before ancestor's end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <small>
|     <small>
|       "a"

#data
<span><strike></span>a
#errors
no DOCTYPE
element not closed before ancestor's end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <strike>
|     <strike>
|       "a"

#data
<span><strile></span>a
#errors
no DOCTYPE
unexpected end tag
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <strile>
|     "a"

#data
<span><strong></span>a
#errors
no DOCTYPE
element not closed before ancestor's end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <strong>
|     <strong>
|       "a"

#data
<span><tt></span>a
#errors
no DOCTYPE
element not closed before ancestor's end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <tt>
|     <tt>
|       "a"

#data
<span><u></span>a
#errors
no DOCTYPE
element not closed before ancestor's end tag
unexpected EOF
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <u>
|     <u>
|       "a"

#data
<span><time></span>a
#errors
no DOCTYPE
element not closed before ancestor's end tag
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <time>
|     "a"

#data
<span><option></span>a
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <option>
|     "a"

#data
<span><optgroup></span>a
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <optgroup>
|     "a"

#data
<span><rp></span>a
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <rp>
|     "a"

#data
<span><rt></span>a
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <rt>
|     "a"

#data
<span><unknown></span>a
#errors
no DOCTYPE
element not closed before ancestor's end tag
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <unknown>
|     "a"

#data
<span><event-source></span>a
#errors
no DOCTYPE
element not closed before ancestor's end tag
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <event-source>
|     "a"

#data
<address><option></address>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <address>
|       <option>
|     "x"

#data
<article><option></article>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <article>
|       <option>
|     "x"

#data
<aside><option></aside>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <aside>
|       <option>
|     "x"

#data
<blockquote><option></blockquote>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <blockquote>
|       <option>
|     "x"

#data
<center><option></center>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <center>
|       <option>
|     "x"

#data
<datagrid><option></datagrid>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <datagrid>
|       <option>
|     "x"

#data
<details><option></details>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <details>
|       <option>
|     "x"

#data
<dialog><option></dialog>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <dialog>
|       <option>
|     "x"

#data
<dir><option></dir>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <dir>
|       <option>
|     "x"

#data
<div><option></div>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <option>
|     "x"

#data
<dl><option></dl>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <dl>
|       <option>
|     "x"

#data
<fieldset><option></fieldset>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <fieldset>
|       <option>
|     "x"

#data
<figure><option></figure>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <figure>
|       <option>
|     "x"

#data
<footer><option></footer>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <footer>
|       <option>
|     "x"

#data
<header><option></header>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <header>
|       <option>
|     "x"

#data
<listing><option></listing>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <listing>
|       <option>
|     "x"

#data
<menu><option></menu>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <menu>
|       <option>
|     "x"

#data
<nav><option></nav>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <nav>
|       <option>
|     "x"

#data
<ol><option></ol>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <ol>
|       <option>
|     "x"

#data
<pre><option></pre>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <pre>
|       <option>
|     "x"

#data
<section><option></section>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <section>
|       <option>
|     "x"

#data
<ul><option></ul>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <ul>
|       <option>
|     "x"

#data
<form><option></form>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <form>
|       <option>
|     "x"

#data
<p><option></p>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <option>
|     "x"

#data
<dd><option></dd>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <dd>
|       <option>
|     "x"

#data
<dt><option></dt>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <dt>
|       <option>
|     "x"

#data
<li><option></li>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <li>
|       <option>
|     "x"

#data
<h3><option></h3>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <h3>
|       <option>
|     "x"

#data
<address><optgroup></address>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <address>
|       <optgroup>
|     "x"

#data
<article><optgroup></article>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <article>
|       <optgroup>
|     "x"

#data
<aside><optgroup></aside>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <aside>
|       <optgroup>
|     "x"

#data
<blockquote><optgroup></blockquote>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <blockquote>
|       <optgroup>
|     "x"

#data
<center><optgroup></center>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <center>
|       <optgroup>
|     "x"

#data
<datagrid><optgroup></datagrid>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <datagrid>
|       <optgroup>
|     "x"

#data
<details><optgroup></details>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <details>
|       <optgroup>
|     "x"

#data
<dialog><optgroup></dialog>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <dialog>
|       <optgroup>
|     "x"

#data
<dir><optgroup></dir>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <dir>
|       <optgroup>
|     "x"

#data
<div><optgroup></div>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <optgroup>
|     "x"

#data
<dl><optgroup></dl>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <dl>
|       <optgroup>
|     "x"

#data
<fieldset><optgroup></fieldset>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <fieldset>
|       <optgroup>
|     "x"

#data
<figure><optgroup></figure>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <figure>
|       <optgroup>
|     "x"

#data
<footer><optgroup></footer>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <footer>
|       <optgroup>
|     "x"

#data
<header><optgroup></header>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <header>
|       <optgroup>
|     "x"

#data
<listing><optgroup></listing>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <listing>
|       <optgroup>
|     "x"

#data
<menu><optgroup></menu>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <menu>
|       <optgroup>
|     "x"

#data
<nav><optgroup></nav>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <nav>
|       <optgroup>
|     "x"

#data
<ol><optgroup></ol>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <ol>
|       <optgroup>
|     "x"

#data
<pre><optgroup></pre>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <pre>
|       <optgroup>
|     "x"

#data
<section><optgroup></section>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <section>
|       <optgroup>
|     "x"

#data
<ul><optgroup></ul>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <ul>
|       <optgroup>
|     "x"

#data
<form><optgroup></form>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <form>
|       <optgroup>
|     "x"

#data
<p><optgroup></p>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <optgroup>
|     "x"

#data
<dt><optgroup></dt>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <dt>
|       <optgroup>
|     "x"

#data
<dd><optgroup></dd>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <dd>
|       <optgroup>
|     "x"

#data
<h3><optgroup></h3>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <h3>
|       <optgroup>
|     "x"

#data
<applet><option></applet>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <applet>
|       <option>
|     "x"

#data
<applet><optgroup></applet>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <applet>
|       <optgroup>
|     "x"

#data
<button><option></button>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <button>
|       <option>
|     "x"

#data
<button><optgroup></button>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <button>
|       <optgroup>
|     "x"

#data
<marquee><option></marquee>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <marquee>
|       <option>
|     "x"

#data
<marquee><optgroup></marquee>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <marquee>
|       <optgroup>
|     "x"

#data
<applet><option></applet>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <applet>
|       <option>
|     "x"

#data
<object><optgroup></object>x
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <object>
|       <optgroup>
|     "x"

#data
<ruby><option><rp></ruby>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <ruby>
|       <option>
|       <rp>

#data
<ruby><option><rt></ruby>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <ruby>
|       <option>
|       <rt>

#data
<ruby><optgroup><rp></ruby>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <ruby>
|       <optgroup>
|       <rp>

#data
<ruby><optgroup><rt></ruby>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <ruby>
|       <optgroup>
|       <rt>

#data
<span><option></span>a
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <option>
|     "a"

#data
<span><optgroup></span>a
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <span>
|       <optgroup>
|     "a"

#data
<table><caption><option></caption></table>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <option>

#data
<table><caption><optgroup></caption></table>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <optgroup>

#data
<table><caption><optgroup><option></caption></table>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <caption>
|         <optgroup>
|           <option>

#data
<table><tr><td><option></td></table>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <option>

#data
<table><tr><td><optgroup></td></table>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <optgroup>

#data
<table><tr><th><option></th></table>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             <option>

#data
<table><tr><th><optgroup></th></table>
#errors
no DOCTYPE
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <th>
|             <optgroup>

#data
<body></caption>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></col>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></colgroup>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></frame>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></frameset>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></head>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></tbody>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></td>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></tfoot>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></th>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></thead>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></tr>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></script>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></select>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></style>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></table>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></textarea>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></title>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></noscript>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></noembed>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></noframes>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<body></iframe>
#errors
no DOCTYPE
unmatched end tag
#document
| <html>
|   <head>
|   <body>

#data
<!DOCTYPE HTML>a<table>b<thead>c</table>
#errors
text in table
x2
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     <table>
|       <thead>

#data
<!DOCTYPE HTML>a<table>b<tbody>c</table>
#errors
text in table
x2
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     <table>
|       <tbody>

#data
<!DOCTYPE HTML>a<table>b<tfoot>c</table>
#errors
text in table
x2
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     <table>
|       <tfoot>

#data
<!DOCTYPE HTML>a<table>b<tr>c</table>
#errors
text in table
x2
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     <table>
|       <tbody>
|         <tr>

#data
<!DOCTYPE HTML>a<table>b<thead>c<tr>d</table>
#errors
text in table
x
3
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     "d"
|     <table>
|       <thead>
|         <tr>

#data
<!DOCTYPE HTML>a<table>b<td></td>c</table>
#errors
text in table
x2
no <tr>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<!DOCTYPE HTML>a<table>b<th></th>c</table>
#errors
text in table
x2
no <tr>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     <table>
|       <tbody>
|         <tr>
|           <th>

#data
<!DOCTYPE HTML>a<table>b<tr>c<td></td>d</table>
#errors
text in table
x
3
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     "d"
|     <table>
|       <tbody>
|         <tr>
|           <td>

#data
<!DOCTYPE HTML>a<table>b<tr>c<th></th>d</table>
#errors
text in table
x
3
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     "d"
|     <table>
|       <tbody>
|         <tr>
|           <th>

#data
<!DOCTYPE HTML>a<table>b<col>c</table>
#errors
text in table
x2
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     <table>
|       <colgroup>
|         <col>

#data
<!DOCTYPE HTML>a<table>b<colgroup>c</table>
#errors
text in table
x2
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     <table>
|       <colgroup>

#data
<!DOCTYPE HTML>a<table>b<caption></caption>c</table>
#errors
text in table
x2
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     <table>
|       <caption>

#data
<!DOCTYPE HTML>a<table>b<thead>c<caption></caption>d</table>
#errors
text in table
x
3
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     "d"
|     <table>
|       <thead>
|       <caption>

#data
<!DOCTYPE HTML>a<table>b<thead>c<col>d</table>
#errors
text in table
x
3
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     "d"
|     <table>
|       <thead>
|       <colgroup>
|         <col>

#data
<!DOCTYPE HTML>a<table>b<thead>c<colgroup>d</table>
#errors
text in table
x
3
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     "d"
|     <table>
|       <thead>
|       <colgroup>

#data
<!DOCTYPE HTML>a<table>b<tr>c<col>d</table>
#errors
text in table
x
3
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     "d"
|     <table>
|       <tbody>
|         <tr>
|       <colgroup>
|         <col>

#data
<!DOCTYPE HTML>a<table>b<tr>c<colgroup>d</table>
#errors
text in table
x
3
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     "d"
|     <table>
|       <tbody>
|         <tr>
|       <colgroup>

#data
<!DOCTYPE HTML>a<table>b<tr>c<caption></caption>d</table>
#errors
text in table
x
3
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     "d"
|     <table>
|       <tbody>
|         <tr>
|       <caption>

#data
<!DOCTYPE HTML>a<table>b<!---->c</table>
#errors
text in table
x2
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     <table>
|       <!--  -->

#data
<!DOCTYPE HTML>a<table>b<style></style>c</table>
#errors
text in table
start tag in table
text in table
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     <style>
|     "c"
|     <table>

#data
<!DOCTYPE HTML>a<table>b<script></script>c</table>
#errors
text in table
start tag in table
text in table
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     <script>
|     "c"
|     <table>

#data
<!DOCTYPE HTML>a<table>b<input type=hidden>c</table>
#errors
text in table
start tag in table
text in table
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     <input>
|       type="hidden"
|     "c"
|     <table>

#data
<!DOCTYPE HTML>a<table><tbody>b<!---->c</table>
#errors
text in table
x2
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     "c"
|     <table>
|       <tbody>
|         <!--  -->

#data
<!DOCTYPE HTML>a<table><tbody>b<style></style>c</table>
#errors
text in table
start tag in table
text in table
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     <style>
|     "c"
|     <table>
|       <tbody>

#data
<!DOCTYPE HTML>a<table><tbody>b<script></script>c</table>
#errors
text in table
start tag in table
text in table
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     <script>
|     "c"
|     <table>
|       <tbody>

#data
<!DOCTYPE HTML>a<table><tbody>b<input type=hidden>c</table>
#errors
text in table
start tag in table
text in table
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     <input>
|       type="hidden"
|     "c"
|     <table>
|       <tbody>

#data
<!DOCTYPE HTML>a<table>b&lt;</table>
#errors
text in table
x2
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b<"
|     <table>

#data
<!DOCTYPE HTML>a<table>b<span>c&lt;</span>d</table>
#errors
text in table
<span> in table
text in table
x2
</span> in table
text in table
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     <span>
|       "c<"
|     "d"
|     <table>

#data
<!DOCTYPE HTML>a<table>b<plaintext>c&lt;</plaintext>d</table>
#errors
text in table
<plaintext> in table
text in table
:
:
:
:
x6
unexpected EOF
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     <plaintext>
|       "c&lt;</plaintext>d</table>"
|     <table>

#data
<!DOCTYPE HTML>a<table>b<xmp>c&lt;<p></xmp>d</table>
#errors
text in table
<xmp> in table
text in table
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     <xmp>
|       "c&lt;<p>"
|     "d"
|     <table>

#data
<!DOCTYPE HTML>a<table>b<title>c&lt;<p></title>d</table>
#errors
text in table
<title> in table
text in table
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   <body>
|     "a"
|     "b"
|     <title>
|       "c<<p>"
|     "d"
|     <table>

#data
<!DOCTYPE HTML></head> <base> a
#errors
metadata element after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <base>
|   " "
|   " "
|   <body>
|     "a"

#data
<!DOCTYPE HTML></head> <link> a
#errors
metadata element after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <link>
|   " "
|   " "
|   <body>
|     "a"

#data
<!DOCTYPE HTML></head> <meta> a
#errors
metadata element after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <meta>
|   " "
|   " "
|   <body>
|     "a"

#data
<!DOCTYPE HTML></head> <noframes></noframes> a
#errors
metadata element after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <noframes>
|   " "
|   " "
|   <body>
|     "a"

#data
<!DOCTYPE HTML></head> <script></script> a
#errors
metadata element after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <script>
|   " "
|   " "
|   <body>
|     "a"

#data
<!DOCTYPE HTML></head> <style></style> a
#errors
metadata element after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <style>
|   " "
|   " "
|   <body>
|     "a"

#data
<!DOCTYPE HTML></head> <title></title> a
#errors
metadata element after </head>
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|     <title>
|   " "
|   " "
|   <body>
|     "a"

#data
<!DOCTYPE HTML></head> &#x20;
#errors
#document
| <!DOCTYPE HTML>
| <html>
|   <head>
|   "  "
|   <body>
